Feat : CPP00 module ex02
The Account.hpp file is present and a quick compilation of tests.cpp confirms that an Account.cpp file is missing. There’s also an old output log that seems to contain the matching output.
So you need to create a Account.cpp file and quickly write some lines of pure awesome C++ and after a couple of failed compilations, your program will compiles and passes the tests with a perfect output, except for the timestamps differences.

 #11